Added alert resolve functionlality.
Alert resolver is run at the end of each rule processing, before rule alert (if any) is sent.
The resolver checks elasticsearch if there was an alert for this rule in the previous run. If there was,
and in this run there are no alerts anymore, it asks rule alerters to resolve the alert.